nlp/requirements.txt
# NLP Pipeline - iyisiniye.com
# Türkçe yemek yorumu analizi için gerekli paketler

# Transformer modelleri (Türkçe BERT)
transformers>=4.36.0

# PyTorch CPU-only
# Kurulum: pip install torch --index-url https://download.pytorch.org/whl/cpu
# NOT: torch ayrı kurulur, --index-url requirements.txt'te kullanılamaz
torch

# Fuzzy string matching (yemek adı eşleştirme)
rapidfuzz>=3.5.0

# Veri işleme
pandas>=2.1.0
numpy>=1.24.0

# Makine öğrenmesi
scikit-learn>=1.3.0

# İlerleme çubuğu
tqdm>=4.66.0

# PostgreSQL bağlantısı
psycopg2-binary>=2.9.9
